requirements.txt
python-dotenv
selenium
chromedriver-autoinstaller
pandas
matplotlib
seaborn
requests
